+++ axes/gl/source/shader.cpp
@@ -21,7 +21,7 @@ StatusOr<Shader> Shader::CompileSource(const char* source, ShaderType type) {
   if (!success) {
     GLint log_length;
     glGetShaderiv(shader_id, GL_INFO_LOG_LENGTH, &log_length);
-    List<GLchar> info_log(log_length);
+    List<GLchar> info_log(static_cast<size_t>(log_length));
     glGetShaderInfoLog(shader_id, log_length, nullptr, info_log.data());
     return utils::InvalidArgumentError(info_log.data());
   }
